refactor: cleanup legacy utils and provenance exports
