+++ msmu/_utils/__init__.py
@@ -1,6 +1,4 @@
-from .subset import split_tmt
-from ._modality import get_modality_dict, get_label, add_quant, reindex_obs
-from .._core._provenance import append_cmd_log, uns_logger
+from ._mudata import reindex_obs
 from .peptide import (
     _calc_exp_mz,
     _count_missed_cleavages,
@@ -11,15 +9,8 @@ from .protein import select_repr_protein
 from .fasta import map_fasta, attach_fasta, parse_uniprot_accession
 
 __all__ = [
-    "subset",
-    "split_tmt",
-    "get_modality_dict",
     "map_fasta",
     "attach_fasta",
-    "get_label",
-    "uns_logger",
-    "append_cmd_log",
-    "add_quant",
     "reindex_obs",
     "_calc_exp_mz",
     "_count_missed_cleavages",
