--- msmu/_utils/_provenance.py
@@ -1,19 +1,0 @@
-"""Backward-compatible provenance helpers re-exported from :mod:`msmu._core`."""
-
-from .._core._provenance import (
-    append_cmd_log,
-    capture_provenance_output,
-    get_bound_call_kwargs,
-    normalize_cmd_for_runtime,
-    serialize,
-    uns_logger,
-)
-
-__all__ = [
-    "append_cmd_log",
-    "capture_provenance_output",
-    "get_bound_call_kwargs",
-    "normalize_cmd_for_runtime",
-    "serialize",
-    "uns_logger",
-]
